+++ src/test/java/nextstep/subway/favorite/FavoriteAcceptanceTest.java
@@ -2,7 +2,7 @@ package nextstep.subway.favorite;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
+import nextstep.subway.utils.AcceptanceTest;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
