+++ src/test/java/nextstep/subway/member/MemberAcceptanceTest.java
@@ -2,7 +2,7 @@ package nextstep.subway.member;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
+import nextstep.subway.utils.AcceptanceTest;
 import nextstep.subway.auth.dto.TokenResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
