+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.path.acceptance;
 import com.google.common.collect.Lists;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
+import nextstep.subway.utils.AcceptanceTest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
